requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
flask_migrate
spotipy
requests
pydub
reportlab
PyJWT
python-dotenv
deezer-python
Flask-Assets
musicbrainzngs
openai
flask-admin
Flask-Login
Flask-Mail
Flask-Session
authlib
Flask-Caching
psutil